+++ http/comment.http
@@ -25,7 +25,7 @@ POST {{baseUrl}}/comments/articles/21
 Content-Type: application/json
 
 {
-  "content": "최애 영화"
+  "content": "웃기면서 사랑스러운 영화"
 }
 ###
 POST {{baseUrl}}/comments/products/29
