+++ http/product.http
@@ -21,7 +21,7 @@ GET {{baseUrl}}/products?name=Pro&description=기능
 
 // 상품 상세 조회, 좋아요, 등록 인증 필요
 ###
-GET {{baseUrl}}/products/24
+GET {{baseUrl}}/products/8
 
 ###
 POST {{baseUrl}}/products/29/like/toggle
@@ -50,11 +50,11 @@ Content-Type: application/json
 }
 
 ###
-PATCH {{baseUrl}}/products/29
+PATCH {{baseUrl}}/products/8
 Content-Type: application/json
 
 {
-  "price": 1300000
+  "price": 120000
 }
 
 ###
