+++ src/router/article.router.ts
@@ -1,16 +1,16 @@
 import express from 'express';
 import articleControl from '../controller/article.control';
-import authenticateUser from '../middleware/authenticate.user';
-import authorizeUser from '../middleware/authorize.user';
+import authenticate from '../middleware/authenticate';
+import authorize from '../middleware/authorize';
 import withTryCatch from '../lib/withTryCatch';
 
 const articleRouter = express.Router();
 
 articleRouter.get('/', withTryCatch(articleControl.getList));
-articleRouter.get('/:id', authenticateUser, withTryCatch(articleControl.get));
-articleRouter.post('/:id/like/toggle', authenticateUser, withTryCatch(articleControl.likeToggle));
-articleRouter.post('/', authenticateUser, withTryCatch(articleControl.post));
-articleRouter.patch('/:id', authenticateUser, authorizeUser, withTryCatch(articleControl.patch));
-articleRouter.delete('/:id', authenticateUser, authorizeUser, withTryCatch(articleControl.erase));
+articleRouter.get('/:id', authenticate, withTryCatch(articleControl.get));
+articleRouter.post('/:id/like/toggle', authenticate, withTryCatch(articleControl.likeToggle));
+articleRouter.post('/', authenticate, withTryCatch(articleControl.post));
+articleRouter.patch('/:id', authenticate, authorize, withTryCatch(articleControl.patch));
+articleRouter.delete('/:id', authenticate, authorize, withTryCatch(articleControl.erase));
 
 export default articleRouter;
