+++ src/router/auth.router.ts
@@ -1,13 +1,13 @@
 import express from 'express';
 import authControl from '../controller/auth.control';
 import withTryCatch from '../lib/withTryCatch';
-import authenticateUser from '../middleware/authenticate.user';
+import authenticate from '../middleware/authenticate';
 
 const authRouter = express.Router();
 
 authRouter.post('/register', withTryCatch(authControl.register));
 authRouter.post('/login', withTryCatch(authControl.login));
-authRouter.post('/logout', authenticateUser, withTryCatch(authControl.logout));
+authRouter.post('/logout', authenticate, withTryCatch(authControl.logout));
 
 // 토큰 재발행
 authRouter.get('/tokens/view', withTryCatch(authControl.viewTokens)); // 토큰 확인: 부가 기능
