+++ src/router/notification.router.ts
@@ -1,12 +1,12 @@
 import express from 'express';
 import notiControl from '../controller/notification.control';
 import withTryCatch from '../lib/withTryCatch';
-import authenticateUser from '../middleware/authenticate.user';
+import authenticate from '../middleware/authenticate';
 
 const notiRouter = express.Router();
 
-notiRouter.get('/', authenticateUser, withTryCatch(notiControl.getList)); // 알림 목록 조회
-notiRouter.get('/:id', authenticateUser, withTryCatch(notiControl.countUnread)); // 알림 조회
-notiRouter.patch('/:id', authenticateUser, withTryCatch(notiControl.patch)); // 알림 수정 (단방향: 읽음으로)
+notiRouter.get('/', authenticate, withTryCatch(notiControl.getList)); // 알림 목록 조회
+notiRouter.get('/:id', authenticate, withTryCatch(notiControl.countUnread)); // 알림 조회
+notiRouter.patch('/:id', authenticate, withTryCatch(notiControl.patch)); // 알림 수정 (단방향: 읽음으로)
 
 export default notiRouter;
