+++ src/service/auth.service.ts
@@ -5,11 +5,11 @@ import { ACCESS_TOKEN_COOKIE_NAME, REFRESH_TOKEN_COOKIE_NAME } from '../lib/cons
 import { generateTokens, verifyRefreshToken } from '../lib/token';
 import NotFoundError from '../middleware/errors/NotFoundError';
 import { assert } from 'superstruct';
-import { CreateUser } from '../struct/userStruct';
+import { CreateUser } from '../struct/user.struct';
 import { Response } from 'express';
-import { CreateUserDto, LoginDto } from '../dto/dto';
+import { CreateUserDto, LoginDto } from '../types/dto';
 import { User } from '@prisma/client';
-import { SafeUser, TokenType } from '../dto/interfaceType';
+import { SafeUser, TokenType } from '../types/interfaceType';
 import { getIO } from '../websocket/socketIO';
 
 async function register(data: CreateUserDto): Promise<SafeUser> {
