+++ src/struct/article.struct.ts
