+++ tsconfig.json
@@ -22,5 +22,5 @@
     /* Completeness */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/**/*", "prisma/mock.ts", "src/seed.ts", "src/websocket/websocket.ts"]
+  "include": ["src/**/*", "src/mock.ts", "src/seed.ts", "src/websocket/websocket.ts", "src/main.ts"]
 }
